Added in current eclipse project
In its current state the game is playable, however there are known bugs
with death events and occasionally broken collisions. The sprites and
images are currently placeholder and will be updated. I anticipate to
use the animation and sound abilities of processing 3.0 to create
animations and sound effects in the near future (after bug fixes).